TASK: Move content graph projection interfaces to `Projection\ContentGraph`
