+++ Neos.ContentRepository.Core/Classes/Projection/ContentGraph/ContentGraphProjectionInterface.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Neos\ContentRepository\Core\Projection;
+namespace Neos\ContentRepository\Core\Projection\ContentGraph;
 
-use Neos\ContentRepository\Core\ContentGraphReadModelInterface;
+use Neos\ContentRepository\Core\Projection\ProjectionInterface;
 
 /**
  * @extends ProjectionInterface<ContentGraphReadModelInterface>
- * @api for creating a custom content repository graph and workspace implementation, **not for users of the CR**
+ * @api for creating a custom content repository graph projection implementation, **not for users of the CR**
  */
 interface ContentGraphProjectionInterface extends ProjectionInterface
 {
