+++ Neos.ContentRepository.Core/Classes/Projection/ProjectionsAndCatchUpHooks.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Neos\ContentRepository\Core\Projection;
 
+use Neos\ContentRepository\Core\Projection\ContentGraph\ContentGraphProjectionInterface;
+
 /**
  * @internal
  */
